+++ Changelog.md
@@ -1,6 +1,7 @@
 # 2026.05.08
 * Build orders are open with new `build_order.html` page.
 * Clean the game JS files to remove the BO content.
+* Add "Submit your build order" button.
 
 # 2026.05.01
 * Script 'update_build_orders' supports `-r` option to remove BO.